Extract plugin compat handling into separate file
